requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.27.0
anthropic>=0.39.0
notion-client>=2.2.0,<2.3.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
mcp>=1.0.0
httpx>=0.27.0
